PHPStan level 2

+++ actions/entire/UsersAction.php
@@ -154,12 +154,12 @@ final class UsersAction extends BaseAction
     {
         $t2 = $_SERVER['REQUEST_TIME'] ?? time();
         $t1 = gmmktime(
-            gmdate('H', $t2),
-            gmdate('i', $t2),
-            gmdate('s', $t2) + 1,
-            gmdate('n', $t2),
-            gmdate('j', $t2) - 1,
-            gmdate('Y', $t2)
+            (int)gmdate('H', $t2),
+            (int)gmdate('i', $t2),
+            (int)gmdate('s', $t2) + 1,
+            (int)gmdate('n', $t2),
+            (int)gmdate('j', $t2) - 1,
+            (int)gmdate('Y', $t2)
         );
         $query = (new Query())
             ->select([
